Add files via upload
In this repository a how to guide in software engineering for novices; it covers software development practises involved in testing and the techniques used. As well as that the guide has a section where the reader develops their programming skills by learning python following exercises in which they can do and follow. 
In completion of the guide this will make the reader go from novice to beginner.